+++ financialaid/models.py
@@ -1,8 +1,6 @@
 """
 Models for the Financial Aid App
 """
-import datetime
-
 from django.contrib.auth.models import User
 from django.core.exceptions import ValidationError
 from django.db import (
@@ -15,52 +13,9 @@ from financialaid.constants import FinancialAidStatus
 from micromasters.models import (
     AuditableModel,
     AuditModel,
+    TimestampedModel,
 )
 from micromasters.utils import serialize_model_object
-
-
-class TimestampedModelQuerySet(models.query.QuerySet):
-    """
-    Subclassed QuerySet for TimestampedModelManager
-    """
-    def update(self, **kwargs):
-        """
-        Automatically update updated_on timestamp when .update(). This is because .update()
-        does not go through .save(), thus will not auto_now, because it happens on the
-        database level without loading objects into memory.
-        """
-        if "updated_on" not in kwargs:
-            kwargs["updated_on"] = datetime.datetime.utcnow()
-        return super().update(**kwargs)
-
-
-class TimestampedModelManager(models.Manager):
-    """
-    Subclassed manager for TimestampedModel
-    """
-    def update(self, **kwargs):
-        """
-        Allows access to TimestampedModelQuerySet's update method on the manager
-        """
-        return self.get_queryset().update(**kwargs)
-
-    def get_queryset(self):
-        """
-        Returns custom queryset
-        """
-        return TimestampedModelQuerySet(self.model, using=self._db)
-
-
-class TimestampedModel(models.Model):
-    """
-    Base model for create/update timestamps
-    """
-    objects = TimestampedModelManager()
-    created_on = models.DateTimeField(auto_now_add=True)  # UTC
-    updated_on = models.DateTimeField(auto_now=True)  # UTC
-
-    class Meta:
-        abstract = True
 
 
 class Tier(TimestampedModel):
