+++ micromasters/models.py
@@ -2,23 +2,68 @@
 Classes related to models for MicroMasters
 """
 
+import datetime
 from django.contrib.auth.models import User
 from django.contrib.postgres.fields import JSONField
 from django.db import transaction
 from django.db.models import (
     DateTimeField,
     ForeignKey,
+    Manager,
     Model,
     SET_NULL,
 )
+from django.db.models.query import QuerySet
 
 
-class AuditModel(Model):
+class TimestampedModelQuerySet(QuerySet):
+    """
+    Subclassed QuerySet for TimestampedModelManager
+    """
+    def update(self, **kwargs):
+        """
+        Automatically update updated_on timestamp when .update(). This is because .update()
+        does not go through .save(), thus will not auto_now, because it happens on the
+        database level without loading objects into memory.
+        """
+        if "updated_on" not in kwargs:
+            kwargs["updated_on"] = datetime.datetime.utcnow()
+        return super().update(**kwargs)
+
+
+class TimestampedModelManager(Manager):
+    """
+    Subclassed manager for TimestampedModel
+    """
+    def update(self, **kwargs):
+        """
+        Allows access to TimestampedModelQuerySet's update method on the manager
+        """
+        return self.get_queryset().update(**kwargs)
+
+    def get_queryset(self):
+        """
+        Returns custom queryset
+        """
+        return TimestampedModelQuerySet(self.model, using=self._db)
+
+
+class TimestampedModel(Model):
+    """
+    Base model for create/update timestamps
+    """
+    objects = TimestampedModelManager()
+    created_on = DateTimeField(auto_now_add=True)  # UTC
+    updated_on = DateTimeField(auto_now=True)  # UTC
+
+    class Meta:
+        abstract = True
+
+
+class AuditModel(TimestampedModel):
     """An abstract base class for audit models"""
 
     acting_user = ForeignKey(User, null=True, on_delete=SET_NULL)
-    created_on = DateTimeField(auto_now_add=True)
-    updated_on = DateTimeField(auto_now=True)
     data_before = JSONField(blank=True, null=True)
     data_after = JSONField(blank=True, null=True)
